[REF] foodhub_custom: stock.pack.operation -> stock.move.line
+ split models into separate files
