+++ foodhub_custom/models/__init__.py
@@ -1,1 +1,3 @@
-from . import stock
+from . import stock_move
+from . import stock_move_line
+from . import stock_picking
